refactor: doamin영역으로 이동
